% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zelenium.R
\docType{data}
\name{list_infos}
\alias{list_infos}
\title{Hold number of infos requested}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
list_infos
}
\description{
Hold number of infos requested
}
\keyword{datasets}
