% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longrunning.R
\name{long_running}
\alias{long_running}
\title{Long running task}
\usage{
long_running(number)
}
\arguments{
\item{number:}{Define number of iterations}
}
\value{
array of numbers
}
\description{
Long running task
}
