% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\docType{data}
\name{pkg.env}
\alias{pkg.env}
\title{Hold number of infos requested}
\format{
An object of class \code{environment} of length 1.
}
\usage{
pkg.env
}
\description{
Hold number of infos requested
}
\keyword{datasets}
